Copy only binary and dynamically linked files to scratch container
This commit reduces the container size drastically by following the
instructions from.

http://blog.oddbit.com/2015/02/05/creating-minimal-docker-images/

In addition I have removed the LICENSE file as FFMPEG is GPL and I'm
unsure of the implications of containerising a GPL codebase and really
couldn't be bothered dealing with this now.

I've also updated the README.md to give credit where credit is due and
updated the FFMPEG version to 4.1.